# Re-use big.Int in DecomposePQ