Old:
```
goos: windows
goarch: amd64
pkg: github.com/gotd/td/internal/crypto
BenchmarkDecomposePQ
BenchmarkDecomposePQ-4   	      16	 707597638 ns/op	33065308 B/op	 3485209 allocs/op
PASS
```

New: 
```
goos: windows
goarch: amd64
pkg: github.com/gotd/td/internal/crypto
BenchmarkDecomposePQ
BenchmarkDecomposePQ-4           26   541487342 ns/op    863167 B/op    107167 allocs/op
PASS
```

benchcmp
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkDecomposePQ-4     707597638     541487342     -23.48%

benchmark                  old allocs     new allocs     delta
BenchmarkDecomposePQ-4     3485209        107167         -96.93%

benchmark                  old bytes     new bytes     delta
BenchmarkDecomposePQ-4     33065308      863167        -97.39%
```